cpp_brown_belt (week_6): Failed case #5/9: Wrong answer invalid token 5.55599e+07 in "... 43 unique stops, 5.55599e+07 route length, 29.5119 ..."

+++ cpp_brown_belt/week_6/database.h
@@ -44,8 +44,8 @@ public:
     std::optional<const Stop*>  GetStopInfo (const std::string& name) const;
     std::pair<double,double> GetBusDistance(const std::string& name) const;
 private:
-    std::pair<double,double> ComputeDistance(const Bus& bus) const;
+    std::pair<double,unsigned int> ComputeDistance(const Bus& bus) const;
     Stops stops_;
     Buses buses_;
-    mutable std::unordered_map<std::string_view,std::pair<double,double>> bus_to_distance_;
+    mutable std::unordered_map<std::string_view,std::pair<double,unsigned int>> bus_to_distance_;
 };
